Moved the Docker executor declaration into a header.
This moves the declaration of the Docker executor into the
Docker executor header file and moves the code for the Docker
executor binary into a new launcher implementation file.

This change will enable the Mesos executor driver
implementation to make use of the `DockerExecutor` symbol.

Review: https://reviews.apache.org/r/71033/
